use current_author method instead of overriding params
