--- app/controllers/concerns/subdomain_as_user_id.rb
@@ -1,11 +1,0 @@
-module SubdomainAsUserId
-  extend ActiveSupport::Concern
-
-  included do
-    before_action :expose_user_id_param!
-  end
-
-  def expose_user_id_param!
-    request.params[:user_id] = request.subdomain
-  end
-end
