--- app/controllers/pages_controller.rb
@@ -1,5 +1,0 @@
-class PagesController < ApplicationController
-  def home
-    redirect_to notes_url
-  end
-end
