+++ app/controllers/users/notes_controller.rb
@@ -1,12 +1,11 @@
 module Users
-  class NotesController < AuthorizedController
+  class NotesController < Users::BaseController
     skip_before_action :authenticate_user!, only: %i[index show]
-    skip_verify_authorized only: %i[index show]
     skip_before_action :authorize_resource!, only: %i[index show]
+    skip_verify_authorized only: %i[index show]
 
     expose_decorated :notes, :filtered_notes
-    expose_decorated :note, find_by: :slug, parent: :user
-    expose :user, find_by: :username
+    expose_decorated :note, find_by: :slug, parent: :current_author
 
     helper_method :filter_params
 
@@ -55,7 +54,7 @@ module Users
     end
 
     def raw_notes
-      user.notes
+      current_author.notes
     end
 
     def filter_params
