+++ app/controllers/users/pages_controller.rb
@@ -1,6 +1,8 @@
 module Users
-  class PagesController < ApplicationController
-    expose_decorated :user, find_by: :username
+  class PagesController < Users::BaseController
+    skip_before_action :authenticate_user!, only: %i[index]
+    skip_before_action :authorize_resource!, only: %i[index]
+    skip_verify_authorized only: %i[index]
 
     def index
     end
