+++ app/views/users/pages/index.html.slim
@@ -1,23 +1,23 @@
 .grid-y
   .cell
-    h2 = title(user.username)
+    h2 = title(current_author.username)
   .cell
-    p = user.email
+    p = current_author.email
   .cell
     .grid-x
       .cell
         p
-          => user.username
+          => current_author.username
           | has
-          =<> link_to "#{user.notes_count} #{'note'.pluralize(user.notes_count)}",
-                      notes_url(subdomain: user.username)
+          =<> link_to "#{current_author.notes_count} #{'note'.pluralize(current_author.notes_count)}",
+                      notes_url(subdomain: current_author.username)
           | with average score ★
-          =< user.average_rating
+          =< current_author.average_rating
       .cell
-        - if current_user&.subscribed_to?(user)
-          = button_to "Unsubscribe", subscription_path(current_user.subscription_for(user)),
+        - if current_user&.subscribed_to?(current_author)
+          = button_to "Unsubscribe", subscription_path(current_user.subscription_for(current_author)),
                       method: :delete, data: { confirm: 'Are you sure?' }, class: "button"
         - else
           = button_to "Subscribe", subscriptions_path, method: :post,
-                      params: { subscription: { followee_id: user.id } },
+                      params: { subscription: { followee_id: current_author.id } },
                       class: "button"
