+++ config/routes.rb
@@ -8,7 +8,7 @@ Rails.application.routes.draw do
     end
   end
 
-  root to: "pages#home"
+  root to: "notes#index"
 
   resources :subscriptions, only: %i[create destroy]
   resources :notes, only: %i[index]
